fixed 1

+++ Temp.java
@@ -6,7 +6,7 @@ public class Temp {
         ch = args[0].charAt(0);
         v = Float.parseFloat(args[1]);
         if (ch == 'f' || ch == 'F') {
-            ans = (v - 32f) * 5 / 9;
+            ans = (float)(v - 32) * 5 / 9;
             System.out.println("Celsius: " + ans);
         } else if (ch == 'c' || ch == 'C') {
             ans = v * 1.8f + 32f;
